+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Show privacy notice when creating a user or user group [#466](https://github.com/openkfw/TruBudget/issues/466)
 - Added search feature for histories [#425](https://github.com/openkfw/TruBudget/issues/425)
 - Changed date format to DD.MM.YYYY [#480](https://github.com/openkfw/TruBudget/issues/480)
+- Documents can be uploaded and are shared via offchain-storage [#481](https://github.com/openkfw/TruBudget/issues/481)
 
 <!-- ### Changed -->